requirements.txt
asgiref==3.8.1
Django==5.0.3
django-rest-framework==0.1.0
djangorestframework==3.15.1
psycopg2==2.9.9
python-dotenv==1.0.1
setuptools==69.1.1
sqlparse==0.4.4
wheel==0.43.0
